pyproject.toml
[project]
name = "ai-prosouze"
version = "0.1.0"
description = ""
authors = [
    {name = "X0re4ik",email = "xoore4ik@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.6.2.post1)",
    "certifi (==2024.8.30)",
    "charset-normalizer (==3.4.0)",
    "click (==8.1.7)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.5)",
    "fastapi-cli (==0.0.5)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.27.2)",
    "idna (==3.10)",
    "jinja2 (==3.1.4)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "pydantic (==2.10.1)",
    "pydantic-core (==2.27.1)",
    "pygments (==2.18.0)",
    "python-dotenv (==1.0.1)",
    "python-multipart (==0.0.17)",
    "pyyaml (==6.0.2)",
    "requests (==2.32.3)",
    "rich (==13.9.4)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.41.3)",
    "typer (==0.13.1)",
    "typing-extensions (==4.12.2)",
    "urllib3 (==2.2.3)",
    "uvicorn (==0.32.1)",
    "uvloop (==0.21.0)",
    "watchfiles (==1.0.0)",
    "websockets (==14.1)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "aiogram (>=3.22.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

